feat: Add support for Laravel Vapor file upload
This commit adds support for uploading files to Laravel Vapor. It includes changes to the `Flows/Commands.vue` component and the `flows.ts` store file. The `uploadFile` method in the store now uses the `laravel-vapor` package to store the file in a public-read-write bucket. The progress of the upload is tracked and displayed in the UI.