feat: Enable NUXT_HUB_BLOB_REMOTE configuration option
This commit adds the NUXT_HUB_BLOB_REMOTE configuration option to the .env.example file and updates the nuxt.config.ts file to use the value of NUXT_HUB_BLOB_REMOTE for the remote property in the hub configuration. This change allows for more flexibility in configuring the handling of blob data in the application.

Note: Please remove any unnecessary details from the commit message before using it.